Refactor KafkaMessageType into model since several apps need it
